perf(web): load the formatting bar's panel lazily to keep the entry chunk under budget
Mounting the bar eagerly in boot.ts pulled the verb table — and with it
@codemirror/state and the icon set — into the critical path: 157.6 KB
gzip against a 138 KB budget (verify's bundle-size gate). The component in
the entry chunk is now only the gate (registry, keyboard state, coarse
pointer); the bar itself is TouchFormattingBarPanel, loaded the first time
the gate opens. Measured after the split: 129.2 KB.
